Break apart artboardAdders from Select Overlay
